% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.R
\name{build_heatmap_data}
\alias{build_heatmap_data}
\title{Heatmap-ready data summarizing location, bin count, and magic.}
\usage{
build_heatmap_data(bucket_df = load_bucket_sweep())
}
\arguments{
\item{bucket_df}{Data frame returned by \code{\link[=load_bucket_sweep]{load_bucket_sweep()}}.}
}
\value{
Tidy data frame.
}
\description{
Heatmap-ready data summarizing location, bin count, and magic.
}
