% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_slices}
\alias{create_slices}
\title{Slice frsr sample data into equal-width bins.}
\usage{
create_slices(df, N, min_input = 0.5, max_input = 2)
}
\arguments{
\item{df}{Data frame containing an \code{input} column.}

\item{N}{Number of slices.}

\item{min_input}{Minimum input to keep.}

\item{max_input}{Maximum input to keep.}
}
\value{
Tidy data frame with \code{slice} column.
}
\description{
Slice frsr sample data into equal-width bins.
}
