% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_optimal_magic}
\alias{find_optimal_magic}
\title{Find the optimal magic constant for a slice.}
\usage{
find_optimal_magic(slice_data)
}
\arguments{
\item{slice_data}{Slice produced by \code{\link[=create_slices]{create_slices()}}.}
}
\value{
List containing the minimum magic and its objective value.
}
\description{
Find the optimal magic constant for a slice.
}
