% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.R
\name{load_bucket_sweep}
\alias{load_bucket_sweep}
\title{Load precomputed bucket sweeps.}
\usage{
load_bucket_sweep(filename = "too_many_bins.csv")
}
\arguments{
\item{filename}{Name of the CSV stored under \code{inst/extdata}.}
}
\value{
Data frame.
}
\description{
Load precomputed bucket sweeps.
}
