% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.R
\name{load_varied_bins}
\alias{load_varied_bins}
\title{Load the varied bin dataset with ordered Bottom factor.}
\usage{
load_varied_bins()
}
\value{
Data frame.
}
\description{
Load the varied bin dataset with ordered Bottom factor.
}
