% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logStratifiedSampler}
\alias{logStratifiedSampler}
\title{Log-stratified sampler for floats.}
\usage{
logStratifiedSampler(min, max, n)
}
\arguments{
\item{min}{Minimum sample value.}

\item{max}{Maximum sample value.}

\item{n}{Number of points to draw.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Log-stratified sampler for floats.
}
