% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstructed.R
\name{prep_deconstruction_df}
\alias{prep_deconstruction_df}
\title{Prepare a deconstruction dataframe for plotting.}
\usage{
prep_deconstruction_df(df, NRmax)
}
\arguments{
\item{df}{Output from \code{\link[frsrr:frsr_sample]{frsrr::frsr_sample()}} with \code{keep_params = TRUE}.}

\item{NRmax}{Newton-Raphson iteration ceiling used during sampling.}
}
\value{
Tidy data frame with derived ranking columns.
}
\description{
Prepare a deconstruction dataframe for plotting.
}
