% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.R
\name{sample_clusters}
\alias{sample_clusters}
\title{Sample representative clusters for downstream visualizations.}
\usage{
sample_clusters(
  n = 16384,
  cluster_ranges = list(c(1597622357, 1597176038), c(1598049921, 1597918602),
    c(1597884263, 1596910052))
)
}
\arguments{
\item{n}{Number of samples per cluster.}

\item{cluster_ranges}{List of length three containing magic ranges.}
}
\value{
Tibble with input/error/magic/cluster columns.
}
\description{
Sample representative clusters for downstream visualizations.
}
