% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualfrsr-package.R
\docType{package}
\name{visualfrsr-package}
\alias{visualfrsr}
\alias{visualfrsr-package}
\title{visualfrsr: Visual Explorations of Fast Reciprocal Square Root Approximations}
\description{
Provides data generators, binning utilities, clustering helpers, and
standardized plotting functions that were previously scattered across scripts
in the FastInverseSqrt-Visualized project. The package exposes those
capabilities as reusable functions so datasets and visualizations can be
reproduced programmatically or integrated into pipelines.
}
\author{
\strong{Maintainer}: Adam Hyland \email{0x5f37642f@protonmail.com}

}
\keyword{internal}
